done deal

+++ parsing/minishell.h
@@ -6,7 +6,7 @@
 /*   By: hznagui <hznagui@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/01 12:17:32 by hznagui           #+#    #+#             */
-/*   Updated: 2023/04/15 13:50:24 by hznagui          ###   ########.fr       */
+/*   Updated: 2023/04/15 16:21:32 by hznagui          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,7 +55,7 @@ char	**ft_split22(t_data *a);
 char	*ft_strdup(char *s1);
 size_t	ft_strlen(char *s);
 char	*ft_substr(char *s, unsigned int start, size_t len);
-char	**free_all22(char **str, size_t max);
+char	**free_all22(char **str, int max);
 char	*ft_strjoin(char *s1, char *s2);
 void	*ft_memcpy(void *dst, void *src, int n);
 int	ft_check_cmd(char *cmd);
